refactor: move git into utils
