+++ libs/cy-reporter/src/utils/git.ts
