+++ libs/cy-reporter/src/utils/index.ts
@@ -2,7 +2,8 @@ import path from 'path'
 import fs from 'fs'
 import { green, red, cyan, bgGreen, bgRed, bgWhite, gray, yellow } from 'chalk'
 import axios from 'axios'
-import { baseURL } from './env'
+import { baseURL } from '../env'
+export * from './git'
 
 export const client = axios.create({ baseURL })
